Datenmanagement_Chapter1
This stuff is way too inefficient. it takes too long to read the chapters and condense the knowledge down to a word document -> the ROI is too low, I probably still can't recall most of the information. THERE ARE REVIEW QUESTIONS AT THE END NOONE SAID ANYTHING ABOUT!! It's probably easier to answer those and actively engage with this stuff rather than trying to sum this all up