Webtech Eigenstudium_D Inkl. Python version :D
I wanted to do the PHP stuff also in python cuz I did a course in the summer at codecademy and I thought it would be such a waste to not use python. funny enough, it is also possible to run Python on the XAMPP Server stuff. At least the Apache worked. It is perhaps possible to also get the other stuff working too. (MySQL n stuff). I'll do both things for this class. it's so cool to see it work :D